philo

+++ philosophers/src/time.c
@@ -1,6 +1,6 @@
 #include "../philo.h"
 
-size_t	get_current_time(void)
+int	get_current_time(void)
 {
 	struct timeval	time;
 
@@ -9,9 +9,9 @@ size_t	get_current_time(void)
 	return (time.tv_sec * 1000 + time.tv_usec / 1000);
 }
 
-int	ft_usleep(size_t milliseconds)
+int	ft_usleep(int milliseconds)
 {
-	size_t	start;
+	int	start;
 
 	start = get_current_time();
 	while ((get_current_time() - start) < milliseconds)
